Move map iterators to their own file
